Migration RegulationConditon model
