--- config/validator/Application/Regulation/Command/Steps/SaveRegulationStep4Command.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping https://symfony.com/schema/dic/constraint-mapping/constraint-mapping-1.0.xsd">
-    <class name="App\Application\Regulation\Command\Steps\SaveRegulationStep4Command">
-        <property name="maxWeight">
-            <constraint name="Positive" />
-        </property>
-        <property name="maxHeight">
-            <constraint name="Positive" />
-        </property>
-        <property name="maxWidth">
-            <constraint name="Positive" />
-        </property>
-        <property name="maxLength">
-            <constraint name="Positive" />
-        </property>
-    </class>
-</constraint-mapping>
