+++ src/Application/Regulation/Query/GetRegulationOrdersToDatexFormatQueryHandler.php
@@ -6,7 +6,6 @@ namespace App\Application\Regulation\Query;
 
 use App\Application\Regulation\View\DatexLocationView;
 use App\Application\Regulation\View\RegulationOrderDatexListItemView;
-use App\Application\Regulation\View\VehicleCharacteristicsView;
 use App\Domain\Regulation\Repository\RegulationOrderRecordRepositoryInterface;
 
 final class GetRegulationOrdersToDatexFormatQueryHandler
@@ -22,11 +21,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandler
         $regulationOrderViews = [];
 
         foreach ($regulationOrders as $regulationOrder) {
-            $hasVehicleCharacteristics = $regulationOrder['maxWeight']
-                || $regulationOrder['maxHeight']
-                || $regulationOrder['maxLength']
-                || $regulationOrder['maxWidth'];
-
             $regulationOrderViews[] = new RegulationOrderDatexListItemView(
                 $regulationOrder['uuid'],
                 $regulationOrder['issuingAuthority'],
@@ -44,12 +38,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandler
                     toLongitude: $regulationOrder['toLongitude'],
                     toLatitude: $regulationOrder['toLatitude'],
                 ),
-                $hasVehicleCharacteristics ? new VehicleCharacteristicsView(
-                    $regulationOrder['maxWeight'],
-                    $regulationOrder['maxHeight'],
-                    $regulationOrder['maxWidth'],
-                    $regulationOrder['maxLength'],
-                ) : null,
             );
         }
