+++ src/Application/Regulation/View/RegulationOrderDatexListItemView.php
@@ -13,7 +13,6 @@ final class RegulationOrderDatexListItemView
         public readonly ?\DateTimeInterface $startDate,
         public readonly ?\DateTimeInterface $endDate,
         public readonly ?DatexLocationView $location,
-        public readonly ?VehicleCharacteristicsView $vehicleCharacteristics,
     ) {
     }
 }
