+++ src/Infrastructure/Controller/Regulation/Steps/Step2Controller.php
@@ -66,7 +66,7 @@ final class Step2Controller extends AbstractRegulationController
 
             if (!$commandFailed) {
                 return new RedirectResponse(
-                    url: $this->router->generate('app_regulations_steps_4', ['uuid' => $uuid]),
+                    url: $this->router->generate('app_regulation_detail', ['uuid' => $uuid]),
                     status: Response::HTTP_SEE_OTHER,
                 );
             }
