+++ src/Infrastructure/Persistence/Doctrine/Mapping/Regulation.Measure.orm.xml
@@ -10,5 +10,6 @@
             <join-column name="location_uuid" referenced-column-name="uuid" nullable="false" on-delete="CASCADE"/>
         </join-columns>
     </many-to-one>
+    <one-to-many field="regulationConditions" target-entity="App\Domain\Condition\RegulationCondition" mapped-by="measure"/>
   </entity>
 </doctrine-mapping>
