+++ templates/regulation/_items.html.twig
@@ -84,7 +84,7 @@
                 </tr>
             {% else %}
                 <tr data-testid="empty-row">
-                    <td colspan="4">
+                    <td colspan="3">
                         <p><em>{{ 'regulation.list.empty'|trans }}</em></p>
                     </td>
                 </tr>
