--- templates/regulation/steps/step4.html.twig
@@ -1,37 +1,0 @@
-{% extends 'regulation/steps/base.html.twig' %}
-
-{% block title %}
-    {{ 'regulation.stepper.steps'|trans({"%step%": 4}) }} - {{ parent() }}
-{% endblock %}
-
-{% block step_body %}
-    {{ form_start(form) }}
-        {{ form_errors(form) }}
-
-        <fieldset class="fr-fieldset">
-            <legend class="fr-fieldset__legend">
-                {{ 'regulation.step4.vehicle_characteristics'|trans }}
-                <span class="fr-hint-text fr-text--sm">
-                    {{ 'regulation.step4.vehicle_characteristics.help'|trans }}
-                </span>
-            </legend>
-            <div class="fr-fieldset__element">
-                {{ form_row(form.maxWeight, {group_class: 'fr-input-group', widget_class: 'fr-input'}) }}
-            </div>
-            <div class="fr-fieldset__element">
-                {{ form_row(form.maxHeight, {group_class: 'fr-input-group', widget_class: 'fr-input'}) }}
-            </div>
-            <div class="fr-fieldset__element">
-                {{ form_row(form.maxWidth, {group_class: 'fr-input-group', widget_class: 'fr-input'}) }}
-            </div>
-            <div class="fr-fieldset__element">
-                {{ form_row(form.maxLength, {group_class: 'fr-input-group', widget_class: 'fr-input'}) }}
-            </div>
-        </fieldset>
-
-        <a href="{{ path('app_regulations_steps_2', {uuid}) }}" class="fr-btn fr-btn--tertiary fr-btn--icon-left fr-icon-arrow-left-line fr-mr-3w">
-            {{ "common.back"|trans }}
-        </a>
-        {{ form_widget(form.save, {attr: {class: 'fr-btn fr-btn--icon-right fr-icon-arrow-right-line'}}) }}
-    {{ form_end(form) }}
-{% endblock %}
