+++ tests/Integration/Infrastructure/Controller/Api/get-regulations-expected-result.xml
@@ -79,27 +79,6 @@
                             </loc:linearWithinLinearElement>
                         </locationByOrder>
                     </conditions>
-                    <conditions xsi:type="VehicleCondition">
-                        <negate>false</negate>
-                        <vehicleCharacteristics>
-                            <com:lengthCharacteristic>
-                                <com:comparisonOperator>greaterThanOrEqualTo</com:comparisonOperator>
-                                <com:vehicleLength>6</com:vehicleLength>
-                            </com:lengthCharacteristic>
-                                <com:heightCharacteristic>
-                                <com:comparisonOperator>greaterThanOrEqualTo</com:comparisonOperator>
-                                <com:vehicleHeight>1.8</com:vehicleHeight>
-                            </com:heightCharacteristic>
-                            <com:widthCharacteristic>
-                                <com:comparisonOperator>greaterThanOrEqualTo</com:comparisonOperator>
-                                <com:vehicleWidth>2</com:vehicleWidth>
-                            </com:widthCharacteristic>
-                            <com:grossWeightCharacteristic>
-                                <com:comparisonOperator>greaterThanOrEqualTo</com:comparisonOperator>
-                                <com:grossVehicleWeight>3.5</com:grossVehicleWeight>
-                            </com:grossWeightCharacteristic>
-                        </vehicleCharacteristics>
-                    </conditions>
                 </condition>
             </trafficRegulation>
         </trafficRegulationOrder>
