+++ tests/Unit/Application/Regulation/Command/DuplicateRegulationCommandHandlerTest.php
@@ -6,17 +6,13 @@ namespace App\Tests\Unit\Application\Regulation\Command;
 
 use App\Application\CommandBusInterface;
 use App\Application\Regulation\Query\Location\GetLocationByRegulationOrderQuery;
-use App\Application\Condition\Query\VehicleCharacteristics\GetVehicleCharacteristicsByRegulationConditionQuery;
 use App\Application\IdFactoryInterface;
 use App\Application\QueryBusInterface;
 use App\Application\Regulation\Command\DuplicateRegulationCommand;
 use App\Application\Regulation\Command\DuplicateRegulationCommandHandler;
 use App\Application\Regulation\Command\Steps\SaveRegulationStep1Command;
-use App\Application\Regulation\Command\Steps\SaveRegulationStep4Command;
 use App\Domain\Regulation\Location;
-use App\Domain\Condition\RegulationCondition;
 use App\Domain\Regulation\Repository\LocationRepositoryInterface;
-use App\Domain\Condition\VehicleCharacteristics;
 use App\Domain\Regulation\Exception\RegulationCannotBeDuplicated;
 use App\Domain\Regulation\RegulationOrder;
 use App\Domain\Regulation\RegulationOrderRecord;
@@ -36,7 +32,6 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
     private $originalRegulationOrderRecord;
     private $organization;
     private $originalRegulationOrder;
-    private $originalRegulationCondition;
 
     public function setUp(): void
     {
@@ -47,12 +42,7 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
         $this->commandBus = $this->createMock(CommandBusInterface::class);
         $this->locationRepository = $this->createMock(LocationRepositoryInterface::class);
         $this->organization = $this->createMock(Organization::class);
-        $this->originalRegulationCondition = $this->createMock(RegulationCondition::class);
         $this->originalRegulationOrder = $this->createMock(RegulationOrder::class);
-        $this->originalRegulationOrder
-            ->expects(self::once())
-            ->method('getRegulationCondition')
-            ->willReturn($this->originalRegulationCondition);
 
         $this->originalRegulationOrderRecord = $this->createMock(RegulationOrderRecord::class);
         $this->originalRegulationOrderRecord
@@ -151,11 +141,6 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
         $step1command->startDate = $startDate;
         $step1command->endDate = $endDate;
 
-        $this->originalRegulationCondition
-            ->expects(self::exactly(1))
-            ->method('getUuid')
-            ->willReturn('d71d7c51-2a4b-49e2-8746-436466db1ade');
-
         $this->originalRegulationOrder
             ->expects(self::once())
             ->method('getUuid')
@@ -214,45 +199,17 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
                 )
             );
 
-        // VehicleCharacteristics
-        $vehicleCharacteristics = $this->createMock(VehicleCharacteristics::class);
-        $vehicleCharacteristics
-            ->expects(self::once())
-            ->method('getMaxHeight')
-            ->willReturn(2.0);
-        $vehicleCharacteristics
-            ->expects(self::once())
-            ->method('getMaxWeight')
-            ->willReturn(3.5);
-        $vehicleCharacteristics
-            ->expects(self::once())
-            ->method('getMaxWidth')
-            ->willReturn(2.0);
-        $vehicleCharacteristics
-            ->expects(self::once())
-            ->method('getMaxLength')
-            ->willReturn(10.0);
-
-        $step4Command = new SaveRegulationStep4Command($duplicatedRegulationOrderRecord);
-        $step4Command->maxHeight = 2.0;
-        $step4Command->maxLength = 10.0;
-        $step4Command->maxWeight = 3.5;
-        $step4Command->maxWidth = 2.0;
-
         $this->queryBus
-            ->expects(self::exactly(2))
+            ->expects(self::exactly(1))
             ->method('handle')
-            ->withConsecutive(
-                [new GetLocationByRegulationOrderQuery('2d2e886c-90a3-4cd5-b0bf-ae288ca45830')],
-                [new GetVehicleCharacteristicsByRegulationConditionQuery('d71d7c51-2a4b-49e2-8746-436466db1ade')],
-            )
-            ->willReturnOnConsecutiveCalls($location, $vehicleCharacteristics);
+            ->with(new GetLocationByRegulationOrderQuery('2d2e886c-90a3-4cd5-b0bf-ae288ca45830'))
+            ->willReturn($location);
 
         $this->commandBus
-            ->expects(self::exactly(2))
+            ->expects(self::exactly(1))
             ->method('handle')
-            ->withConsecutive([$step1command], [$step4Command])
-            ->willReturnOnConsecutiveCalls($duplicatedRegulationOrderRecord);
+            ->with($step1command)
+            ->willReturn($duplicatedRegulationOrderRecord);
 
         $handler = new DuplicateRegulationCommandHandler(
             $this->idFactory,
@@ -320,11 +277,6 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
         $step1command->startDate = $startDate;
         $step1command->endDate = $endDate;
 
-        $this->originalRegulationCondition
-            ->expects(self::exactly(1))
-            ->method('getUuid')
-            ->willReturn('d71d7c51-2a4b-49e2-8746-436466db1ade');
-
         $this->originalRegulationOrder
             ->expects(self::once())
             ->method('getUuid')
@@ -335,19 +287,16 @@ final class DuplicateRegulationCommandHandlerTest extends TestCase
             ->method('save');
 
         $this->queryBus
-            ->expects(self::exactly(2))
+            ->expects(self::exactly(1))
             ->method('handle')
-            ->withConsecutive(
-                [new GetLocationByRegulationOrderQuery('2d2e886c-90a3-4cd5-b0bf-ae288ca45830')],
-                [new GetVehicleCharacteristicsByRegulationConditionQuery('d71d7c51-2a4b-49e2-8746-436466db1ade')],
-            )
-            ->willReturnOnConsecutiveCalls(null, null);
+            ->with(new GetLocationByRegulationOrderQuery('2d2e886c-90a3-4cd5-b0bf-ae288ca45830'))
+            ->willReturn(null);
 
         $this->commandBus
             ->expects(self::exactly(1))
             ->method('handle')
-            ->withConsecutive([$step1command])
-            ->willReturnOnConsecutiveCalls($duplicatedRegulationOrderRecord);
+            ->with($step1command)
+            ->willReturn($duplicatedRegulationOrderRecord);
 
         $handler = new DuplicateRegulationCommandHandler(
             $this->idFactory,
