+++ tests/Unit/Application/Regulation/Query/GetRegulationOrdersToDatexFormatQueryHandlerTest.php
@@ -8,7 +8,6 @@ use App\Application\Regulation\Query\GetRegulationOrdersToDatexFormatQueryHandle
 use App\Application\Regulation\Query\GetRegulationOrdersToDatexFormatQuery;
 use App\Application\Regulation\View\DatexLocationView;
 use App\Application\Regulation\View\RegulationOrderDatexListItemView;
-use App\Application\Regulation\View\VehicleCharacteristicsView;
 use App\Domain\Regulation\Repository\RegulationOrderRecordRepositoryInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -60,10 +59,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandlerTest extends TestCase
             'toHouseNumber' => $location1->toHouseNumber,
             'toLatitude' => $location1->toLatitude,
             'toLongitude' => $location1->toLongitude,
-            'maxWeight' => 3.5,
-            'maxHeight' => 3,
-            'maxWidth' => 2,
-            'maxLength' => 10,
         ];
         $regulationOrder2 = [
             'uuid' => '247edaa2-58d1-43de-9d33-9753bf6f4d30',
@@ -80,10 +75,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandlerTest extends TestCase
             'toHouseNumber' => $location2->toHouseNumber,
             'toLatitude' => $location2->toLatitude,
             'toLongitude' => $location2->toLongitude,
-            'maxWeight' => null,
-            'maxHeight' => null,
-            'maxWidth' => null,
-            'maxLength' => null,
         ];
 
         $regulationOrderRecordRepository
@@ -103,7 +94,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandlerTest extends TestCase
                     startDate: $startDate1,
                     endDate: $endDate1,
                     location: $location1,
-                    vehicleCharacteristics: new VehicleCharacteristicsView(3.5, 3, 2, 10),
                 ),
                 new RegulationOrderDatexListItemView(
                     uuid: '247edaa2-58d1-43de-9d33-9753bf6f4d30',
@@ -112,7 +102,6 @@ final class GetRegulationOrdersToDatexFormatQueryHandlerTest extends TestCase
                     startDate: $startDate2,
                     endDate: null,
                     location: $location2,
-                    vehicleCharacteristics: null,
                 ),
             ],
             $regulationOrders,
